+++ lib/database/api/src/commonMain/kotlin/com/retro99/database/api/books/BookEntity.kt
@@ -36,7 +36,7 @@ interface SeriesWithPositionEntity {
     val uuid: String
     val name: String
     val featured: Int?
-    val position: Int?
+    val position: Double?
     val createdAt: String?
     val updatedAt: String?
 }
